+++ appGUI/preferences/PreferencesUIManager.py
@@ -286,6 +286,7 @@ class PreferencesUIManager:
 
             "tools_iso_tool_shape":     self.ui.plugin_eng_pref_form.tools_iso_group.tool_shape_combo,  # "C1"
             "tools_iso_passes":         self.ui.plugin_eng_pref_form.tools_iso_group.passes_entry,
+            "tools_iso_pad_passes":     self.ui.plugin_eng_pref_form.tools_iso_group.pad_passes_entry,
             "tools_iso_overlap":        self.ui.plugin_eng_pref_form.tools_iso_group.overlap_entry,
             "tools_iso_milling_type":   self.ui.plugin_eng_pref_form.tools_iso_group.milling_type_radio,
             "tools_iso_isotype":        self.ui.plugin_eng_pref_form.tools_iso_group.iso_type_radio,