+++ defaults.py
@@ -355,6 +355,7 @@ class FlatCAMDefaults:
 
         "tools_iso_tool_shape": 0,  # "C1"
         "tools_iso_passes": 1,
+        "tools_iso_pad_passes": 0,
         "tools_iso_overlap": 10,
         "tools_iso_milling_type": "cl",
         "tools_iso_isotype": "full",